requirements-channels.txt
Django>=2.2,<2.3
wagtail>=2.5,<2.12
graphene-django==2.7.1
wagtailmedia
graphql-core==2.2.1
factory-boy==2.12.0
wagtail-factories==2.0.0
django-cors-headers==3.4.0
wagtail-headless-preview==0.1.4

# Channels-specific requirements
channels==1.1.8
asgi_redis
graphql_ws
